moved timeSince into utils
